;ing long

+++ P34FilesDownloader/DownloadProgress.h
@@ -11,9 +11,9 @@
 @interface DownloadProgress : NSObject
 
 @property(nonatomic) CGFloat progress;
-@property(nonatomic) NSInteger downloadedBytes;
-@property(nonatomic) NSInteger totalBytes;
-@property(nonatomic) NSInteger bytesPerSecond;
+@property(nonatomic) unsigned long long downloadedBytes;
+@property(nonatomic) unsigned long long totalBytes;
+@property(nonatomic) unsigned long long bytesPerSecond;
 @property(nonatomic) NSTimeInterval remainingTime;
 
 - (NSString *)remainingTimeString;
